+++ OwlOSC/Test/Test.cs
@@ -5,8 +5,7 @@ namespace OwlOSC{
     public class Test {
 
         void Main(){
-            OscPacket packet = new OscMessage("/",1);
-            if(packet.IsBundle){}
+
         }
         
     }